+++ packages/lib-react-components/package.json
@@ -31,7 +31,8 @@
     "react-dom": "^16.4.1",
     "react-select": "^1.0.0-rc.10",
     "react-swipe": "^5.1.1",
-    "styled-components": "^3.2.6"
+    "styled-components": "^3.2.6",
+    "styled-theming": "^2.2.0"
   },
   "devDependencies": {
     "@storybook/addon-actions": "^3.4.8",
