+++ packages/lib-react-components/src/components/layout/ZooFooter/index.js
@@ -1,0 +1,1 @@
+export { default } from './ZooFooter'
