+++ packages/lib-react-components/src/index.js
@@ -9,7 +9,7 @@ export { default as DisplayNameSlugEditor } from './components/display-name-slug
 export { default as ImageSelector } from './components/image-selector';
 export { default as Paginator } from './components/paginator';
 
-export { default as ZooFooter } from './components/layout/zoo-footer';
+export { default as ZooFooter } from './components/layout/ZooFooter';
 export { default as AdminCheckbox } from './components/layout/admin-checkbox';
 export { default as AdminLayoutIndicator } from './components/layout/admin-layout-indicator';
